+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 ### Added
  - Open output folder button
  - Cap output scale based on model scale
+### Changed 
+ - UI Tweaks
 ### Fixed
  - ROCm showing cuda on front end.
+### Removed
+ - GIMM.
+ - locking the app to 1 instance.
 # RVE 2.3.7
 ### Added:
  - PyTorch 2.9